+++ CHANGELOG.md
@@ -5,6 +5,9 @@
   - Organization invite emails are now sent regardless of server setup
 - **Admin Changes**
   - Added CreatedAt to forum export for topics and replies
+- **Organization Changes**
+  - Unreleased datasets are now displayed for organization viewers on the
+    organization page
 
 ## 47.0.0 (July 16, 2020)